+++ app/src/main/java/ly/count/android/demo/ActivityExampleDeviceId.java
@@ -1,6 +1,7 @@
 package ly.count.android.demo;
 
 import android.app.Activity;
+import android.content.res.Configuration;
 import android.os.Bundle;
 import android.view.View;
 
@@ -58,6 +59,10 @@ public class ActivityExampleDeviceId extends Activity {
         super.onStop();
     }
 
-
+    @Override
+    public void onConfigurationChanged (Configuration newConfig){
+        super.onConfigurationChanged(newConfig);
+        Countly.sharedInstance().onConfigurationChanged(newConfig);
+    }
 
 }