+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   * Adding functionality for filtering crashes
   * Adding functionality to set segmentation for automatic and manual views
   * Adding functionality to ignore activities for automatic view tracking
+  * Adding functinality to track orientation changes
   * Adding call to record a event with a given timestamp
   * Adding option to set tamper protection salt in config object
   * Adding calls to give and remove consent to all features at the same time